+++ src/Util.js
@@ -764,6 +764,12 @@
       }
       return retObj;
     },
+    trimRight: function(str) {
+      return str.replace(/\s+$/, "");
+    },
+    trimLeft: function(str) {
+      return str.replace(/^\s+/, "");
+    },
     /**
      * check intersection of two client rectangles
      * @method